%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  
% 
% on Mon Jul 18 15:08:24 2005.
% 
% Generator was the Rdoc class, which is part of the R.oo package written
% by Henrik Bengtsson, 2001-2004.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{argsToString.Rdoc}
\alias{argsToString.Rdoc}
\alias{Rdoc.argsToString}
\alias{argsToString.Rdoc}
\alias{argsToString,Rdoc-method}


\title{Gets the arguments signature of a function}

\description{
  Gets the arguments signature of a function.
}

\synopsis{argsToString.Rdoc(static, fcn, ...)}
\usage{Rdoc$argsToString(fcn, ...)}

\arguments{
 \item{fcn}{A \code{\link[base]{function}}.}
 \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{list}} of \code{\link[base]{character}} string.
}

\author{Henrik Bengtsson \url{http://www.braju.com/R/}}

\seealso{
  For more information see \code{\link{Rdoc}}.
}


\keyword{internal}
\keyword{methods}
\keyword{documentation}
