%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  
% 
% on Tue May 03 13:27:57 2005.
% 
% Generator was the Rdoc class, which is part of the R.oo package written
% by Henrik Bengtsson, 2001-2004.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{as.character.BasicObject}
\alias{as.character.BasicObject}
\alias{BasicObject.as.character}
\alias{as.character.BasicObject}
\alias{as.character,BasicObject-method}
\keyword{methods}
\keyword{internal}


\title{Gets a character string representing the object}

\description{
 Gets a character string representing the object.
}

\usage{\method{as.character}{BasicObject}(this, ...)}

\value{
 Returns a \code{\link[base]{character}} string representation of the object.
}

\author{Henrik Bengtsson \url{http://www.braju.com/R/}}

\seealso{
  For more information see \code{\link{BasicObject}}.
}

\keyword{programming}
\keyword{methods}
