%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  Interface.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{as.character.Interface}
\alias{as.character.Interface}
\alias{Interface.as.character}
\alias{as.character,Interface-method}

\title{Gets a character string representing the Interface}

\description{
 Gets a character string representing the Interface.
}

\usage{
\method{as.character}{Interface}(x, ...)
}

\value{
 Returns a \code{\link[base]{character}} string.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{Interface}}.
}

\keyword{programming}
\keyword{methods}
\keyword{internal}
\keyword{methods}
