%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  
% 
% on Mon Jul 18 15:08:14 2005.
% 
% Generator was the Rdoc class, which is part of the R.oo package written
% by Henrik Bengtsson, 2001-2004.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{as.character.RccViolationException}
\alias{as.character.RccViolationException}
\alias{RccViolationException.as.character}
\alias{as.character.RccViolationException}
\alias{as.character,RccViolationException-method}


\title{Gets a string representing of the RCC violation}

\description{
 Gets a string representing of the RCC violation of format "[{POSIX date string}] {class name}: {msg}, cf. http://www.maths.lth.se/help/R/RCC/".
}

\usage{\method{as.character}{RccViolationException}(this, ...)}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{character}} string.
}

\examples{\dontrun{For a complete example see help(Exception).}}

\author{Henrik Bengtsson \url{http://www.braju.com/R/}}

\seealso{
  For more information see \code{\link{RccViolationException}}.
}

\keyword{programming}
\keyword{methods}
\keyword{error}
\keyword{internal}
\keyword{methods}
