%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  
% 
% on Tue Feb 15 13:45:38 2005.
% 
% Generator was the Rdoc class, which is part of the R.oo package written
% by Henrik Bengtsson, 2001-2004.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{as.character.RdocException}
\alias{as.character.RdocException}
\alias{RdocException.as.character}
\alias{as.character.RdocException}
\alias{as.character,RdocException-method}
\keyword{methods}
\keyword{internal}


\title{Gets a character string representing of the RdocException}

\description{
 .
}

\usage{\method{as.character}{RdocException}(this, ...)}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{character}} string.
}

\author{Henrik Bengtsson \url{http://www.braju.com/R/}}

\seealso{
  For more information see \code{\link{RdocException}}.
}

\keyword{programming}
\keyword{methods}
\keyword{error}
