%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  
% 
% on Tue May 03 13:27:49 2005.
% 
% Generator was the Rdoc class, which is part of the R.oo package written
% by Henrik Bengtsson, 2001-2004.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{ASCII}
\alias{ASCII}

\alias{ASCII.BEL}
\alias{ASCII.BS}
\alias{ASCII.HT}
\alias{ASCII.LF}
\alias{ASCII.FF}
\alias{ASCII.CR}
\alias{ASCII.SO}
\alias{ASCII.SI}
\alias{ASCII.DC1}
\alias{ASCII.DC3}
\alias{ASCII.ESC}

\title{8-bit ASCII table}

\description{
  ASCII is the 8-bit ASCII table with ASCII characters from 0-255.
}

\examples{
  ch <- ASCII[65+1];  # ch == "A"
}

\author{Henrik Bengtsson \url{http://www.braju.com/R/}}

\seealso{
  \code{\link{charToInt}}()
  \code{\link{intToChar}}()
}

\keyword{character}

\keyword{internal}
