%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  BasicObject.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{BasicObject}
\docType{class}
\alias{BasicObject}


\title{A root class like Object but without references}

\description{
 R.oo\cr
 \bold{Class BasicObject}\cr

 public class \bold{BasicObject}\cr
}

\usage{
BasicObject(core=NULL)
}

\arguments{
  \item{core}{The core value of the object.}
}

\section{Fields and Methods}{
 \bold{Methods:}\cr
\tabular{rll}{
 \tab \code{$} \tab  -\cr
 \tab \code{$<-} \tab  -\cr
 \tab \code{.DollarNames} \tab  -\cr
 \tab \code{.subset2Internal} \tab  -\cr
 \tab \code{[[} \tab  -\cr
 \tab \code{[[<-} \tab  -\cr
 \tab \code{as.character} \tab  -\cr
 \tab \code{attach} \tab  -\cr
 \tab \code{detach} \tab  -\cr
 \tab \code{equals} \tab  -\cr
 \tab \code{extend} \tab  -\cr
 \tab \code{getFields} \tab  -\cr
 \tab \code{getInstantiationTime} \tab  -\cr
 \tab \code{hasField} \tab  -\cr
 \tab \code{hashCode} \tab  -\cr
 \tab \code{isReferable} \tab  -\cr
 \tab \code{newInstance} \tab  -\cr
 \tab \code{objectSize} \tab  -\cr
 \tab \code{print} \tab  -\cr
}


 \bold{Methods inherited from logical}:\cr
Ops,nonStructure,vector-method, Ops,structure,vector-method, Ops,vector,nonStructure-method, Ops,vector,structure-method, as.data.frame, as.raster, coerce,ANY,logical-method


}

\author{Henrik Bengtsson}

\keyword{programming}
\keyword{methods}
\keyword{internal}
\keyword{classes}
