%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  Interface.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{extend.Interface}
\alias{extend.Interface}
\alias{Interface.extend}
\alias{extend.Interface}
\alias{extend,Interface-method}


\title{Extends another Interface class}

\description{
 Extends another Interface class.
}

\usage{
\method{extend}{Interface}(this, ...className, ...)
}

\arguments{
  \item{...className}{The name of new interface.}
  \item{...}{Named values representing the fields of the new instance.}
}

\value{
 Returns an Interface of class \code{className}.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{Interface}}.
}

\keyword{programming}
\keyword{methods}
\keyword{internal}
\keyword{methods}
