%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  
% 
% on Tue Feb 15 13:45:35 2005.
% 
% Generator was the Rdoc class, which is part of the R.oo package written
% by Henrik Bengtsson, 2001-2004.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getClassS4Usage.Rdoc}
\alias{getClassS4Usage.Rdoc}
\alias{Rdoc.getClassS4Usage}
\alias{getClassS4Usage.Rdoc}
\alias{getClassS4Usage,Rdoc-method}
\keyword{methods}
\keyword{internal}


\title{Gets the usage of a S4 class}

\description{
  Gets the usage of a S4 class.
}

\synopsis{getClassS4Usage.Rdoc(static, class, ...)}
\usage{Rdoc$getClassS4Usage(class, ...)}

\arguments{
  \item{class}{A class name (\code{\link[base]{character}} string).}
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{character}} string.
}

\author{Henrik Bengtsson \url{http://www.braju.com/R/}}

\seealso{
  For more information see \code{\link{Rdoc}}.
}

\keyword{documentation}

