%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  getConstructorS3.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getConstructorS3}
\alias{getConstructorS3.default}
\alias{getConstructorS3}

\title{Get a constructor method}

\description{
 Get a constructor method.
}

\usage{
\method{getConstructorS3}{default}(name, ...)
}

\arguments{
  \item{name}{The name of the constructor function.}
  \item{...}{Not used.}
}

\seealso{
  \code{\link{setConstructorS3}}().
  \code{\link[R.methodsS3]{getMethodS3}}.
  \code{\link[R.methodsS3]{isGenericS3}}.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}



\keyword{programming}
\keyword{methods}
