%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  
% 
% on Tue Feb 15 13:45:17 2005.
% 
% Generator was the Rdoc class, which is part of the R.oo package written
% by Henrik Bengtsson, 2001-2004.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getDataPath.Package}
\alias{getDataPath.Package}
\alias{Package.getDataPath}
\alias{getDataPath.Package}
\alias{getDataPath,Package-method}
\keyword{methods}
\keyword{internal}

\alias{Package.getData}
\alias{getData.Package}
\alias{getData,Package-method}


\title{Gets the path to the data (data/) directory of this package}

\description{
  Gets the path to the data (data/) directory of this package.
}

\usage{\method{getDataPath}{Package}(this, ...)}

\arguments{
  \item{...}{Not used.}
}

\value{
  Returns a \code{\link[base]{character}} string.
}

\author{Henrik Bengtsson \url{http://www.braju.com/R/}}

\seealso{
  For more information see \code{\link{Package}}.
}
