%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  
% 
% on Tue Feb 15 13:44:46 2005.
% 
% Generator was the Rdoc class, which is part of the R.oo package written
% by Henrik Bengtsson, 2001-2004.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getDetails.Class}
\alias{getDetails.Class}
\alias{Class.getDetails}
\alias{getDetails.Class}
\alias{getDetails,Class-method}
\keyword{methods}
\keyword{internal}


\title{Lists the fields and methods of a class}

\description{
  Lists the fields and methods of a class (or an object).
}

\usage{\method{getDetails}{Class}(this, private=FALSE, ...)}

\arguments{
  \item{...}{Not used.}
}

\value{
  Returns an invisible \code{\link[base]{character}} string of the class information.
}

\examples{
  getDetails(Exception)
}

\author{Henrik Bengtsson \url{http://www.braju.com/R/}}

\keyword{programming}

\keyword{methods}

