%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  050.Object.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getFieldModifiers.Object}
\alias{getFieldModifiers.Object}
\alias{Object.getFieldModifiers}
\alias{getFieldModifiers,Object-method}
\alias{Object.getFieldModifier}
\alias{getFieldModifier.Object}
\alias{getFieldModifier,Object-method}


\title{Gets all types of field modifiers}

\description{
 Gets all types of field modifiers, if any.
}

\usage{
\method{getFieldModifiers}{Object}(this, ...)
}

\arguments{
  \item{...}{Not used.}
}

\value{
  Returns a named \code{\link[base]{list}}.
}

\author{Henrik Bengtsson}



\keyword{internal}
\keyword{methods}
\keyword{programming}
