%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  Interface.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getFields.Interface}
\alias{getFields.Interface}
\alias{Interface.getFields}
\alias{getFields,Interface-method}

\title{Returns NULL}

\description{
  Returns NULL.
}

\usage{
\method{getFields}{Interface}(...)
}

\arguments{
  \item{...}{Ignored.}
}

\value{
 Always returns \code{\link[base]{NULL}}.
}

\author{Henrik Bengtsson}

\seealso{
  For more information see \code{\link{Interface}}.
}

\keyword{programming}
\keyword{methods}
\keyword{internal}
\keyword{methods}
