%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  
% 
% on Mon Apr 03 17:20:13 2006.
% 
% Generator was the Rdoc class, which is part of the R.oo package written
% by Henrik Bengtsson, 2001-2006.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%



\name{getKnownSubclasses.classRepresentation}
\alias{getKnownSubclasses.classRepresentation}
\alias{getKnownSubclasses.classRepresentation}



\title{Gets the known subclasses}

\description{
  Gets the known subclasses.
}

\usage{\method{getKnownSubclasses}{classRepresentation}(this, ...)}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{character}} string.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  \code{\link[R.oo:setNameFormat.classRepresentation]{*setNameFormat}()}
}


\keyword{methods}
\keyword{internal}
\keyword{documentation}
