%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  
% 
% on Mon Apr 03 17:20:46 2006.
% 
% Generator was the Rdoc class, which is part of the R.oo package written
% by Henrik Bengtsson, 2001-2006.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getManPath.Rdoc}
\alias{getManPath.Rdoc}
\alias{Rdoc.getManPath}
\alias{getManPath.Rdoc}
\alias{getManPath,Rdoc-method}


\title{Gets the path to the directory where the Rd files will be saved}

\description{
  Gets the path to the directory where the Rd files will be saved.
}

\synopsis{getManPath.Rdoc(this, ...)}
\usage{Rdoc$getManPath(...)}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{character}} string.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  \code{\link[R.oo:setManPath.Rdoc]{*setManPath}()}
  For more information see \code{\link{Rdoc}}.
}


\keyword{internal}
\keyword{methods}
\keyword{documentation}
