%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  
% 
% on Tue Feb 15 13:44:59 2005.
% 
% Generator was the Rdoc class, which is part of the R.oo package written
% by Henrik Bengtsson, 2001-2004.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getRdDeclaration.classRepresentation}
\alias{getRdDeclaration.classRepresentation}
\alias{classRepresentation.getRdDeclaration}
\alias{getRdDeclaration.classRepresentation}
\alias{getRdDeclaration,classRepresentation-method}
\keyword{methods}
\keyword{internal}


\title{Gets the class declaration in Rd format}

\description{
  Gets the class declaration in Rd format.
}

\usage{\method{getRdDeclaration}{classRepresentation}(this, ...)}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{character}} string.
}

\author{Henrik Bengtsson \url{http://www.braju.com/R/}}

\seealso{
}

\keyword{documentation}

