%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  
% 
% on Mon Apr 03 17:20:49 2006.
% 
% Generator was the Rdoc class, which is part of the R.oo package written
% by Henrik Bengtsson, 2001-2006.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getSource.RdocException}
\alias{getSource.RdocException}
\alias{RdocException.getSource}
\alias{getSource.RdocException}
\alias{getSource,RdocException-method}


\title{Gets the source of the exception}

\description{
 .
}

\usage{\method{getSource}{RdocException}(this, ...)}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns the source.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{RdocException}}.
}

\keyword{programming}
\keyword{methods}
\keyword{error}
\keyword{internal}
\keyword{methods}
