%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  
% 
% on Tue Feb 15 13:45:06 2005.
% 
% Generator was the Rdoc class, which is part of the R.oo package written
% by Henrik Bengtsson, 2001-2004.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getStackTraceString.Exception}
\alias{getStackTraceString.Exception}
\alias{Exception.getStackTraceString}
\alias{getStackTraceString.Exception}
\alias{getStackTraceString,Exception-method}
\keyword{methods}
\keyword{internal}


\title{Gets the stack trace as a string}

\description{
 .
}

\usage{\method{getStackTraceString}{Exception}(this, ...)}

\arguments{
  \item{...}{Not used.}
}

\value{
  Returns a \code{\link[base]{character}} string.
}

\seealso{
  \code{\link[R.oo:getStackTrace.Exception]{*getStackTrace}()}.
  For more information see \code{\link{Exception}}.
}

\author{Henrik Bengtsson \url{http://www.braju.com/R/}}

\keyword{programming}
\keyword{methods}
\keyword{error}
