%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  
% 
% on Tue May 03 13:27:42 2005.
% 
% Generator was the Rdoc class, which is part of the R.oo package written
% by Henrik Bengtsson, 2001-2004.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getStaticInstance.Class}
\alias{getStaticInstance.Class}
\alias{Class.getStaticInstance}
\alias{getStaticInstance.Class}
\alias{getStaticInstance,Class-method}
\keyword{methods}
\keyword{internal}


\title{Gets the static instance of this class}

\description{
  Gets the static instance of this class.
}

\usage{\method{getStaticInstance}{Class}(this, ...)}

\arguments{
  \item{...}{Not used.}
}

\value{
  Returns a reference to an \code{\link{Object}}.
}

\examples{
  obj <- getStaticInstance(Object)
}

\author{Henrik Bengtsson \url{http://www.braju.com/R/}}

\seealso{
  For more information see \code{\link{Class}}.
}

\keyword{programming}

\keyword{methods}

