%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  
% 
% on Mon Apr 03 17:20:46 2006.
% 
% Generator was the Rdoc class, which is part of the R.oo package written
% by Henrik Bengtsson, 2001-2006.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getUsage.Rdoc}
\alias{getUsage.Rdoc}
\alias{Rdoc.getUsage}
\alias{getUsage.Rdoc}
\alias{getUsage,Rdoc-method}


\title{Gets the usage of a method}

\description{
  Gets the usage of a method.
}

\synopsis{getUsage.Rdoc(static, method, class=NULL, ...)}
\usage{Rdoc$getUsage(method, class=NULL, ...)}

\arguments{
  \item{method}{A method name (\code{\link[base]{character}} string).}
  \item{class}{An optional class name (\code{\link[base]{character}} string).}
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{character}} string.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{Rdoc}}.
}


\keyword{internal}
\keyword{methods}
\keyword{documentation}
