%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  
% 
% on Mon Apr 03 17:20:20 2006.
% 
% Generator was the Rdoc class, which is part of the R.oo package written
% by Henrik Bengtsson, 2001-2006.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getWhen.Exception}
\alias{getWhen.Exception}
\alias{Exception.getWhen}
\alias{getWhen.Exception}
\alias{getWhen,Exception-method}


\title{Gets the time when the Exception was created}

\description{
 Gets the time, as a POSIX object, when the Exception was created.
}

\usage{\method{getWhen}{Exception}(this, ...)}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a POSIX time object.
}

\examples{\dontrun{For a complete example see help(Exception).}}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{Exception}}.
}

\keyword{programming}
\keyword{methods}
\keyword{error}
\keyword{internal}
\keyword{methods}
