%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  
% 
% on Tue May 03 13:28:00 2005.
% 
% Generator was the Rdoc class, which is part of the R.oo package written
% by Henrik Bengtsson, 2001-2004.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{hasField.BasicObject}
\alias{hasField.BasicObject}
\alias{BasicObject.hasField}
\alias{hasField.BasicObject}
\alias{hasField,BasicObject-method}
\keyword{methods}
\keyword{internal}


\title{Checks if a field exists or not}

\description{
  Checks if a field exists or not.
}

\usage{\method{hasField}{BasicObject}(this, field, ...)}

\arguments{
  \item{field}{\code{\link[base]{vector}} of fields to be checked if they exists or not.}
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{logical}} \code{\link[base]{vector}} indicating for each field if it exists or not.
}

\author{Henrik Bengtsson \url{http://www.braju.com/R/}}

\seealso{
  To get the fields of an Object, see \code{\link[R.oo:getFields.BasicObject]{*getFields}()}.
  For more extensive information about the fields in an Object see
  \code{\link[R.oo:ll.BasicObject]{*ll}()}.
  For more information see \code{\link{BasicObject}}.
}

\keyword{programming}
\keyword{methods}
