%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  
% 
% on Tue Feb 15 13:44:53 2005.
% 
% Generator was the Rdoc class, which is part of the R.oo package written
% by Henrik Bengtsson, 2001-2004.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{hashCode.BasicObject}
\alias{hashCode.BasicObject}
\alias{BasicObject.hashCode}
\alias{hashCode.BasicObject}
\alias{hashCode,BasicObject-method}
\keyword{methods}
\keyword{internal}


\title{Gets a hash code for the object}

\description{
 Gets a hash code for the object.
}

\usage{\method{hashCode}{BasicObject}(this, ...)}

\arguments{
  \item{...}{Not used.}
}

\value{
  Returns an \code{\link[base]{integer}}.
}

\author{Henrik Bengtsson \url{http://www.braju.com/R/}}

\seealso{
  \code{\link[R.oo:equals.BasicObject]{*equals}()}
  For more information see \code{\link{BasicObject}}.
}

\keyword{programming}
\keyword{methods}
