%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  Rdoc.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{Rdoc$isKeyword}
\alias{Rdoc$isKeyword}
\alias{isKeyword.Rdoc}
\alias{Rdoc.isKeyword}
\alias{isKeyword,Rdoc-method}

\title{Checks if a word is a Rd keyword}

\description{
  Checks if a word is a Rd keyword.
}

\usage{
## Static method (use this):
## Rdoc$isKeyword(word, ...)

## Don't use the below:
\method{isKeyword}{Rdoc}(this, word, ...)
}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{logical}}.
}

\author{Henrik Bengtsson}

\seealso{
  \code{\link[R.oo:getKeywords.Rdoc]{*getKeywords}()}
  For more information see \code{\link{Rdoc}}.
}


\keyword{internal}
\keyword{methods}
\keyword{documentation}
