%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  
% 
% on Mon Jul 18 15:07:36 2005.
% 
% Generator was the Rdoc class, which is part of the R.oo package written
% by Henrik Bengtsson, 2001-2004.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{isReferable.BasicObject}
\alias{isReferable.BasicObject}
\alias{BasicObject.isReferable}
\alias{isReferable.BasicObject}
\alias{isReferable,BasicObject-method}


\title{Checks if the object is referable or not}

\description{
 Checks if the object is referable or not.
}

\usage{\method{isReferable}{BasicObject}(this, ...)}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{logical}} value, which by default is \code{\link[base:logical]{TRUE}} for all
 \code{\link{BasicObject}}'s.
}

\author{Henrik Bengtsson \url{http://www.braju.com/R/}}

\seealso{
  For more information see \code{\link{BasicObject}}.
}

\keyword{programming}
\keyword{methods}
\keyword{internal}
\keyword{methods}
