%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  
% 
% on Tue May 03 13:29:03 2005.
% 
% Generator was the Rdoc class, which is part of the R.oo package written
% by Henrik Bengtsson, 2001-2004.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{methodsInheritedFrom.Rdoc}
\alias{methodsInheritedFrom.Rdoc}
\alias{Rdoc.methodsInheritedFrom}
\alias{methodsInheritedFrom.Rdoc}
\alias{methodsInheritedFrom,Rdoc-method}
\keyword{methods}
\keyword{internal}


\title{Gets all methods inherited from a class in Rd format}

\description{
  Gets all methods inherited from a class in Rd format.
}

\synopsis{methodsInheritedFrom.Rdoc(this, class, visibility=c("public", "protected", "private"), showDeprecated=FALSE, inheritedFrom=NULL, sort=TRUE, trial=FALSE, ...)}
\usage{Rdoc$methodsInheritedFrom(class, visibility=c("public", "protected", "private"), showDeprecated=FALSE, inheritedFrom=NULL, sort=TRUE, trial=FALSE, ...)}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{character}} string.
}

\author{Henrik Bengtsson \url{http://www.braju.com/R/}}

\seealso{
  \code{\link[R.oo:getRdHierarchy.Rdoc]{*getRdHierarchy}()}
  For more information see \code{\link{Rdoc}}.
}

\keyword{documentation}

