%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  BasicObject.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{newInstance.BasicObject}
\alias{newInstance.BasicObject}
\alias{BasicObject.newInstance}
\alias{newInstance.BasicObject}
\alias{newInstance,BasicObject-method}


\title{Creates a new instance of the same class as this object}

\description{
 Creates a new instance of the same class as this object.
}

\usage{\method{newInstance}{BasicObject}(this, ...)}

\arguments{
  \item{...}{Arguments passed to the constructor of the corresponding
    \code{\link{BasicObject}} class.}
}

\value{
  Returns a reference to an instance of \code{\link{BasicObject}} or a subclass thereof.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  \code{\link{newInstance.Object}}().
  \code{\link{newInstance.Class}}().
  For more information see \code{\link{BasicObject}}.
}



\keyword{internal}
\keyword{methods}
\keyword{programming}
