%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  050.Object.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{newInstance.Object}
\alias{newInstance.Object}
\alias{Object.newInstance}
\alias{newInstance,Object-method}

\title{Creates a new instance of the same class as this object}

\description{
 Creates a new instance of the same class as this object.
}

\usage{
\method{newInstance}{Object}(this, ...)
}

\arguments{
  \item{...}{Arguments passed to the constructor of the corresponding
    \code{\link{Object}} class.}
}

\value{
  Returns a reference to an instance of \code{\link{Object}} or a subclass thereof.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  \code{\link{newInstance.Class}}().
  \code{\link{newInstance.BasicObject}}().
  For more information see \code{\link{Object}}.
}



\keyword{internal}
\keyword{methods}
\keyword{programming}
