%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  
% 
% on Tue May 03 13:27:48 2005.
% 
% Generator was the Rdoc class, which is part of the R.oo package written
% by Henrik Bengtsson, 2001-2004.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{Non-documented objects}
\alias{Non-documented objects}
\keyword{documentation}
\title{Non-documented objects}


\alias{appendVarArgs}
\alias{appendVarArgs.function}
\alias{argsToString}
\alias{attach}
\alias{attach.default}
\alias{callSuperMethodS3}
\alias{callSuperMethodS3.default}
\alias{check}
\alias{clone}
\alias{compile}
\alias{createManPath}
\alias{createName}
\alias{declaration}
\alias{detach}
\alias{detach.default}
\alias{escapeRdFilename}
\alias{extend}
\alias{finalize}
\alias{forName}
\alias{getAuthor}
\alias{getBundle}
\alias{getBundlePackages}
\alias{getCall}
\alias{getClass}
\alias{getClass.default}
\alias{getClasses}
\alias{getClasses.default}
\alias{getClassS4Usage}
\alias{getContents}
\alias{getContribUrl}
\alias{getData}
\alias{getDataPath}
\alias{getDate}
\alias{getDescription}
\alias{getDescriptionFile}
\alias{getDetails}
\alias{getDevelUrl}
\alias{getDocPath}
\alias{getEnvironment}
\alias{getExamplePath}
\alias{getFields}
\alias{getHistory}
\alias{getHowToCite}
\alias{getInstanciationTime}
\alias{getInternalAddress}
\alias{getKeywords}
\alias{getKnownSubclasses}
\alias{getLastException}
\alias{getLicense}
\alias{getMaintainer}
\alias{getManPath}
\alias{getMessage}
\alias{getMethods}
\alias{getMethods.default}
\alias{getName}
\alias{getNameFormat}
\alias{getPackage}
\alias{getPackageNameOf}
\alias{getPath}
\alias{getPosition}
\alias{getRccUrl}
\alias{getRdDeclaration}
\alias{getRdHierarchy}
\alias{getRdMethods}
\alias{getRdTitle}
\alias{getSource}
\alias{getStackTrace}
\alias{getStackTraceString}
\alias{getStaticInstance}
\alias{getSuperclasses}
\alias{getTitle}
\alias{getUrl}
\alias{getUsage}
\alias{getVersion}
\alias{getWhen}
\alias{hasField}
\alias{hasVarArgs}
\alias{hasVarArgs.function}
\alias{hierarchy}
\alias{isAbstract}
\alias{isDeprecated}
\alias{isKeyword}
\alias{isLoaded}
\alias{isPrivate}
\alias{isProtected}
\alias{isPublic}
\alias{isReferable}
\alias{isStatic}
\alias{load}
\alias{load.default}
\alias{methodsInheritedFrom}
\alias{newInstance}
\alias{novirtual}
\alias{objectSize}
\alias{printStackTrace}
\alias{save}
\alias{save.default}
\alias{setClassS3}
\alias{setClassS3.default}
\alias{setManPath}
\alias{setNameFormat}
\alias{showAndWait}
\alias{showContents}
\alias{showDescriptionFile}
\alias{showHistory}
\alias{showHowToCite}
\alias{staticCode}
\alias{unload}

\alias{getClass}
\alias{getClass.Object}
\alias{getClass.BasicObject}
\alias{getClass.default}
\alias{setClassS3}
\alias{setClassS3.default}

\description{
  This page contains aliases for all "non-documented" objects that
  \code{R CMD check} detects in this package.

  Almost all of them are \emph{generic} functions that have specific
  document for the corresponding method coupled to a specific class.
  Other functions are re-defined by \code{setMethodS3()} to
  \emph{default} methods. Neither of these two classes are non-documented
  in reality.
  The rest are deprecated methods.
}

\author{Henrik Bengtsson \url{http://www.braju.com/R/}}

\keyword{internal}

