%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  objectSize.default.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{objectSize}
\alias{objectSize.default}
\alias{objectSize}

\title{Gets the size of the object in bytes}

\description{
  Gets the size of the object in bytes.
  This method is just a wrapper for \code{\link[utils]{object.size}}.
}

\usage{\method{objectSize}{default}(...)}

\arguments{
  \item{...}{Arguments passed to \code{\link[utils]{object.size}}.}
}

\value{
 Returns an \code{\link[base]{integer}}.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  Internally \code{\link[utils]{object.size}}.
}

\keyword{attribute}
\keyword{utilities}

