%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  BasicObject.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{print.BasicObject}
\alias{print.BasicObject}
\alias{BasicObject.print}
\alias{print,BasicObject-method}

\title{Prints an BasicObject}

\description{
 For all objects of class \code{\link{BasicObject}}, this method will print the
 value of \code{as.character()} of the object. Note that this function is
 not called if the argument is not an object of class \code{\link{BasicObject}}.
}

\usage{
\method{print}{BasicObject}(x, ...)
}

\arguments{
  \item{...}{Not used.}
}

\author{Henrik Bengtsson}

\seealso{
  \code{\link[base]{print.default}}()
  For more information see \code{\link{BasicObject}}.
}

\keyword{programming}
\keyword{methods}
\keyword{internal}
\keyword{methods}
