%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  
% 
% on Tue Feb 15 13:44:41 2005.
% 
% Generator was the Rdoc class, which is part of the R.oo package written
% by Henrik Bengtsson, 2001-2004.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{print.Class}
\alias{print.Class}
\alias{Class.print}
\alias{print.Class}
\alias{print,Class-method}
\keyword{methods}
\keyword{internal}


\title{Prints detailed information about the class and its fields and methods}

\description{
 Prints detailed information about the class and its fields and methods.
}

\usage{\method{print}{Class}(x, ...)}

\arguments{
  \item{...}{Not used.}
}

\value{
  Returns nothing.
}

\examples{
  print(Object)
}

\author{Henrik Bengtsson \url{http://www.braju.com/R/}}

\seealso{
  For more information see \code{\link{Class}}.
}

\keyword{programming}

\keyword{methods}

