%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  
% 
% on Tue Feb 15 13:45:04 2005.
% 
% Generator was the Rdoc class, which is part of the R.oo package written
% by Henrik Bengtsson, 2001-2004.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{print.Exception}
\alias{print.Exception}
\alias{Exception.print}
\alias{print.Exception}
\alias{print,Exception-method}
\keyword{methods}
\keyword{internal}


\title{Prints the Exception}

\description{
 . By default the \code{as.character()} representation plus
 the stack trace is printed.
}

\usage{\method{print}{Exception}(x, ...)}

\arguments{
  \item{...}{Not used.}
}

\value{Returns nothing.}

\examples{\dontrun{For a complete example see help(Exception).}}

\author{Henrik Bengtsson \url{http://www.braju.com/R/}}

\seealso{
  \code{\link[R.oo:as.character.Exception]{*as.character}()}.
  \code{\link[R.oo:getStackTrace.Exception]{*getStackTrace}()}.
  \code{\link[R.oo:printStackTrace.Exception]{*printStackTrace}()}.
  For more information see \code{\link{Exception}}.
}

\keyword{programming}
\keyword{methods}
\keyword{error}
