%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  Interface.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{print.Interface}
\alias{print.Interface}
\alias{Interface.print}
\alias{print,Interface-method}

\title{Prints an Interface}

\description{
 For all objects of class \code{\link{Interface}}, this method will print the
 value of \code{as.character()} of the object.
}

\usage{
\method{print}{Interface}(x, ...)
}

\arguments{
  \item{...}{Not used.}
}

\author{Henrik Bengtsson}

\seealso{
  For more information see \code{\link{Interface}}.
}

\keyword{programming}
\keyword{methods}
\keyword{internal}
\keyword{methods}
