%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  
% 
% on Tue Feb 15 13:45:06 2005.
% 
% Generator was the Rdoc class, which is part of the R.oo package written
% by Henrik Bengtsson, 2001-2004.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{printStackTrace.Exception}
\alias{printStackTrace.Exception}
\alias{Exception.printStackTrace}
\alias{printStackTrace.Exception}
\alias{printStackTrace,Exception-method}
\keyword{methods}
\keyword{internal}


\title{Prints the stack trace saved when the exception was created}

\description{
 .
}

\usage{\method{printStackTrace}{Exception}(this, ...)}

\arguments{
  \item{...}{Not used.}
}

\value{
  Returns nothing.
}

\examples{\dontrun{For a complete example see help(Exception).}}

\seealso{
  \code{\link[R.oo:getStackTrace.Exception]{*getStackTrace}()}.
  \code{\link[base]{tryCatch}}().
  For more information see \code{\link{Exception}}.
}

\author{Henrik Bengtsson \url{http://www.braju.com/R/}}

\keyword{programming}
\keyword{methods}
\keyword{error}
