%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  
% 
% on Mon Apr 03 18:46:46 2006.
% 
% Generator was the Rdoc class, which is part of the R.oo package written
% by Henrik Bengtsson, 2001-2006.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{R.KEYWORDS}
\alias{R.KEYWORDS}

\title{Reserved words in R not to be used for object names}

\description{
 Reserved words in R not to be used for object names. \code{R.KEYWORDS} is a \code{\link[base]{character}} \code{\link[base]{vector}} of all reserved
 words in \R according to [1].
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\references{
 [1] R Language Definition, version 2.0.1 (2004-11-27).
}



\keyword{programming}
\keyword{internal}
