%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  RccViolationException.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{RccViolationException}
\docType{class}
\alias{RccViolationException}


 \title{An RccViolationException indicates a violation of the R Coding Conventions (RCC)}

 \description{
  Package:  R.oo \cr
\bold{Class RccViolationException}\cr

\code{\link[R.oo]{Object}}\cr
\code{~~|}\cr
\code{~~+--}\code{try-error}\cr
\code{~~~~~~~|}\cr
\code{~~~~~~~+--}\code{condition}\cr
\code{~~~~~~~~~~~~|}\cr
\code{~~~~~~~~~~~~+--}\code{error}\cr
\code{~~~~~~~~~~~~~~~~~|}\cr
\code{~~~~~~~~~~~~~~~~~+--}\code{simpleError}\cr
\code{~~~~~~~~~~~~~~~~~~~~~~|}\cr
\code{~~~~~~~~~~~~~~~~~~~~~~+--}\code{\link[R.oo]{Exception}}\cr
\code{~~~~~~~~~~~~~~~~~~~~~~~~~~~|}\cr
\code{~~~~~~~~~~~~~~~~~~~~~~~~~~~+--}\code{RccViolationException}\cr

\bold{Directly known subclasses:}\cr
\cr

public static class \bold{RccViolationException}\cr
extends \link[R.oo]{Exception}\cr



  An RccViolationException indicates a violation of the R Coding Conventions (RCC).
  It is generated by \code{setConstructorS3()} and \code{setMethodS3()}.
  It is \emph{not} meant to be caught, but instead the source code that
  violates the RCC should be fixed. For more information about RCC, see
  references below.
 }

 \usage{RccViolationException(...)}

 \arguments{
   \item{...}{Any arguments accepted by the constructor of Exception, i.e.
     one or several \code{\link[base]{character}} strings, which will be concatenated and
     contain informative message about why the RCC was violated.}
 }

 \section{Fields and Methods}{
  \bold{Methods:}\cr
\tabular{rll}{
 \tab \code{\link[R.oo:as.character.RccViolationException]{as.character}} \tab Gets a string representing of the RCC violation.\cr
 \tab \code{\link[R.oo:getRccUrl.RccViolationException]{getRccUrl}} \tab Static method to get a URL where the RCC can be found.\cr
}


 \bold{Methods inherited from Exception}:\cr
as.character, getCall, getLastException, getMessage, getStackTrace, getWhen, print, printStackTrace, throw

 \bold{Methods inherited from error}:\cr
as.character, throw

 \bold{Methods inherited from condition}:\cr
as.character, conditionCall, conditionMessage, print

 \bold{Methods inherited from Object}:\cr
$, $<-, [[, [[<-, as.character, attach, attachLocally, clearCache, clone, detach, equals, extend, finalize, gc, getEnvironment, getFields, getInstantiationTime, getStaticInstance, hasField, hashCode, ll, load, objectSize, print, save

 }

 \details{
   Since it is not possible to assert that the RCC is followed during the
   parsing of the source code, but first only when the source code is
   actually executed.
 }

 \examples{
  \dontrun{
   setConstructorS3("myClass", function() { extends(Object(), .value=0) })
   setMethodS3("MyMethod", "myClass", function(this) { "Hullo!" })
  }
 }

 \author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

 \seealso{
   See also \code{\link[base]{try}}() and \code{\link[base:conditions]{tryCatch}()}.
   For detailed information about exceptions see \code{\link{Exception}}.
   The R Coding Conventions (RCC) can be found at
   \url{http://www.maths.lth.se/help/R/RCC/}.
 }

 \keyword{programming}
 \keyword{methods}
 \keyword{error}
\keyword{classes}
