%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  
% 
% on Tue May 03 13:29:10 2005.
% 
% Generator was the Rdoc class, which is part of the R.oo package written
% by Henrik Bengtsson, 2001-2004.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{reportBug}
\alias{reportBug.default}
\alias{reportBug}

\title{Send a bug report about the last Exception thrown}

\description{
 Send a bug report about the last Exception thrown.

 \bold{NOTE: This method is not supported and does NOT work (yet).}
}

\usage{\method{reportBug}{default}(...)}

\arguments{
  \item{...}{Any arguments accepted by
        \code{\link{InternalErrorException.reportBug}}.}
}

\value{
 Returns nothing.
}

\author{Henrik Bengtsson \url{http://www.braju.com/R/}}

\seealso{
  \code{\link{InternalErrorException.reportBug}}.
}

\keyword{programming}
\keyword{methods}
\keyword{error}
\keyword{internal}
