%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  
% 
% on Tue Feb 15 13:45:16 2005.
% 
% Generator was the Rdoc class, which is part of the R.oo package written
% by Henrik Bengtsson, 2001-2004.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{showDescriptionFile.Package}
\alias{showDescriptionFile.Package}
\alias{Package.showDescriptionFile}
\alias{showDescriptionFile.Package}
\alias{showDescriptionFile,Package-method}
\keyword{methods}
\keyword{internal}


\title{Show the DESCRIPTION file of this package}

\description{
  Show the DESCRIPTION file of this package.
  If the \code{DESCRIPTION} file does not exist, an exception is thrown.
}

\usage{\method{showDescriptionFile}{Package}(this, ...)}

\arguments{
  \item{...}{Not used.}
}

\value{
  Returns nothing.
}

\author{Henrik Bengtsson \url{http://www.braju.com/R/}}

\seealso{
  For more information see \code{\link{Package}}.
}
