%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  Interface.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{uses.Interface}
\alias{uses.Interface}
\alias{Interface.uses}
\alias{uses,Interface-method}
\alias{uses}
\alias{uses.character}

\title{Specifies that an object uses this Interface}

\description{
 Specifies that an object uses this Interface.
}

\usage{
\method{uses}{Interface}(this, ...)
}

\value{
 Returns a \code{\link[base]{character}} \code{\link[base]{vector}} of class names of Interface:s.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{Interface}}.
}

\keyword{programming}
\keyword{methods}
\keyword{internal}
\keyword{methods}
