%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  Rdoc.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{argsToString.Rdoc}
\alias{argsToString.Rdoc}
\alias{Rdoc.argsToString}
\alias{argsToString.Rdoc}
\alias{argsToString,Rdoc-method}


\title{Gets the arguments signature of a function}

\description{
  Gets the arguments signature of a function.
}

\usage{
## Static method (use this):
## Rdoc$argsToString(fcn, ...)

## Don't use the below:
\method{argsToString}{Rdoc}(static, fcn, ...)
}

\arguments{
 \item{fcn}{A \code{\link[base]{function}}.}
 \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{list}} of \code{\link[base]{character}} string.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{Rdoc}}.
}


\keyword{internal}
\keyword{methods}
\keyword{documentation}
