%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  BasicObject.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{as.character.BasicObject}
\alias{as.character.BasicObject}
\alias{BasicObject.as.character}
\alias{as.character,BasicObject-method}

\title{Gets a character string representing the object}

\description{
 Gets a character string representing the object.
}

\usage{
\method{as.character}{BasicObject}(x, ...)
}

\value{
 Returns a \code{\link[base]{character}} string representation of the object.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{BasicObject}}.
}

\keyword{programming}
\keyword{methods}
\keyword{internal}
\keyword{methods}
