%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  
% 
% on Tue May 03 13:27:35 2005.
% 
% Generator was the Rdoc class, which is part of the R.oo package written
% by Henrik Bengtsson, 2001-2004.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{as.character.Class}
\alias{as.character.Class}
\alias{Class.as.character}
\alias{as.character.Class}
\alias{as.character,Class-method}
\keyword{methods}
\keyword{internal}


\title{Returns a short string describing the class}

\description{
 Returns a short string describing the class.
}

\usage{\method{as.character}{Class}(this, ...)}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{character}} string.
}

\examples{
  print(as.character(Object))

  # gives:  "Class Object: no fields, 8 methods (no inherited)"
}

\author{Henrik Bengtsson \url{http://www.braju.com/R/}}

\seealso{
  For more information see \code{\link{Class}}.
}

\keyword{programming}

\keyword{methods}

