%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  ASCII.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{ASCII}
\alias{ASCII}

\alias{ASCII.BEL}
\alias{ASCII.BS}
\alias{ASCII.HT}
\alias{ASCII.LF}
\alias{ASCII.FF}
\alias{ASCII.CR}
\alias{ASCII.SO}
\alias{ASCII.SI}
\alias{ASCII.DC1}
\alias{ASCII.DC3}
\alias{ASCII.ESC}

\title{8-bit ASCII table}

\description{
  ASCII is the 8-bit ASCII table with ASCII characters from 0-255.
}

\examples{
  ch <- ASCII[65+1];  # ch == "A"
}

\author{Henrik Bengtsson}

\seealso{
  \code{\link{charToInt}}()
  \code{\link{intToChar}}()
}



\keyword{character}
\keyword{internal}
