%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  
% 
% on Tue May 03 13:29:02 2005.
% 
% Generator was the Rdoc class, which is part of the R.oo package written
% by Henrik Bengtsson, 2001-2004.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{declaration.Rdoc}
\alias{declaration.Rdoc}
\alias{Rdoc.declaration}
\alias{declaration.Rdoc}
\alias{declaration,Rdoc-method}
\keyword{methods}
\keyword{internal}


\title{Gets the class declaration}

\description{
  Gets the class declaration.
}

\synopsis{declaration.Rdoc(this, class, ...)}
\usage{Rdoc$declaration(class, ...)}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{character}} string.
}

\author{Henrik Bengtsson \url{http://www.braju.com/R/}}

\seealso{
  \code{\link[R.oo:getRdHierarchy.Rdoc]{*getRdHierarchy}()}
  For more information see \code{\link{Rdoc}}.
}

\keyword{documentation}

