%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  Rdoc.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{Rdoc$declaration}
\alias{Rdoc$declaration}
\alias{declaration.Rdoc}
\alias{Rdoc.declaration}
\alias{declaration,Rdoc-method}

\title{Gets the class declaration}

\description{
  Gets the class declaration.
}

\usage{
## Static method (use this):
## Rdoc$declaration(class, ...)

## Don't use the below:
\method{declaration}{Rdoc}(this, class, ...)
}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{character}} string.
}

\author{Henrik Bengtsson}

\seealso{
  For more information see \code{\link{Rdoc}}.
}


\keyword{internal}
\keyword{methods}
\keyword{documentation}
