%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  
% 
% on Tue Feb 15 13:45:33 2005.
% 
% Generator was the Rdoc class, which is part of the R.oo package written
% by Henrik Bengtsson, 2001-2004.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{escapeRdFilename.Rdoc}
\alias{escapeRdFilename.Rdoc}
\alias{Rdoc.escapeRdFilename}
\alias{escapeRdFilename.Rdoc}
\alias{escapeRdFilename,Rdoc-method}
\keyword{methods}
\keyword{internal}


\title{Escape non-valid characters in a filename}

\description{
  Escape non-valid characters in a filename.
}

\synopsis{escapeRdFilename.Rdoc(static, filename, ...)}
\usage{Rdoc$escapeRdFilename(filename, ...)}

\arguments{
  \item{filename}{A filename (\code{\link[base]{character}} string) to be escaped.}
  \item{...}{Not used.}
}

\value{
 Returns \code{\link[base]{character}} string.
}

\author{Henrik Bengtsson \url{http://www.braju.com/R/}}

\seealso{
  For more information see \code{\link{Rdoc}}.
}

\keyword{documentation}

