%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  Rdoc.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{Rdoc$escapeRdFilename}
\alias{Rdoc$escapeRdFilename}
\alias{escapeRdFilename.Rdoc}
\alias{Rdoc.escapeRdFilename}
\alias{escapeRdFilename,Rdoc-method}

\title{Escape non-valid characters in a filename}

\description{
  Escape non-valid characters in a filename.
}

\usage{
## Static method (use this):
## Rdoc$escapeRdFilename(filename, ...)

## Don't use the below:
\method{escapeRdFilename}{Rdoc}(static, filename, ...)
}

\arguments{
  \item{filename}{A filename (\code{\link[base]{character}} string) to be escaped.}
  \item{...}{Not used.}
}

\value{
 Returns \code{\link[base]{character}} string.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{Rdoc}}.
}


\keyword{internal}
\keyword{methods}
\keyword{documentation}
