%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  Rdoc.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getClassS4Usage.Rdoc}
\alias{getClassS4Usage.Rdoc}
\alias{Rdoc.getClassS4Usage}
\alias{getClassS4Usage.Rdoc}
\alias{getClassS4Usage,Rdoc-method}


\title{Gets the usage of a S4 class}

\description{
  Gets the usage of a S4 class.
}

\usage{
## Static method (use this):
## Rdoc$getClassS4Usage(class, ...)

## Don't use the below:
\method{getClassS4Usage}{Rdoc}(static, class, ...)
}

\arguments{
  \item{class}{A class name (\code{\link[base]{character}} string).}
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{character}} string.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{Rdoc}}.
}


\keyword{internal}
\keyword{methods}
\keyword{documentation}
