%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  060.Class.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getDetails.Class}
\alias{getDetails.Class}
\alias{Class.getDetails}
\alias{getDetails,Class-method}

\title{Lists the fields and methods of a class}

\description{
  Lists the fields and methods of a class (or an object).
}

\usage{
\method{getDetails}{Class}(this, private=FALSE, ...)
}

\arguments{
  \item{...}{Not used.}
}

\value{
  Returns an invisible \code{\link[base]{character}} string of the class information.
}

\examples{
  getDetails(Exception)
}

\author{Henrik Bengtsson}



\keyword{internal}
\keyword{methods}
\keyword{programming}
