%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  BasicObject.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getInstanciationTime.BasicObject}
\alias{getInstanciationTime.BasicObject}
\alias{BasicObject.getInstanciationTime}
\alias{getInstanciationTime.BasicObject}
\alias{getInstanciationTime,BasicObject-method}


\title{Gets the time when the object was instanciated}

\description{
 Gets the time when the object was instanciated (created) as a POSIXt object.
}

\usage{\method{getInstanciationTime}{BasicObject}(this, ...)}

\arguments{
  \item{...}{Not used.}
}

\value{
  Returns a POSIXt object, which extends class POSIXct.
}

\seealso{
  For more about time formats and POSIX see \code{\link[base]{DateTimeClasses}}.
  For more information see \code{\link{BasicObject}}.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\keyword{programming}
\keyword{methods}
\keyword{internal}
\keyword{methods}
