%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  
% 
% on Tue Feb 15 13:45:31 2005.
% 
% Generator was the Rdoc class, which is part of the R.oo package written
% by Henrik Bengtsson, 2001-2004.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getKeywords.Rdoc}
\alias{getKeywords.Rdoc}
\alias{Rdoc.getKeywords}
\alias{getKeywords.Rdoc}
\alias{getKeywords,Rdoc-method}
\keyword{methods}
\keyword{internal}


\title{Gets the keywords defined in R with descriptions}

\description{
  Gets the keywords defined in R with descriptions.
}

\synopsis{getKeywords.Rdoc(this, fullInfo=FALSE, ...)}
\usage{Rdoc$getKeywords(fullInfo=FALSE, ...)}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{character}} string.
}

\author{Henrik Bengtsson \url{http://www.braju.com/R/}}

\seealso{
  \code{\link[R.oo:setManPath.Rdoc]{*setManPath}()}
  For more information see \code{\link{Rdoc}}.
}

\keyword{documentation}

