%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  Rdoc.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{Rdoc$getKeywords}
\alias{Rdoc$getKeywords}
\alias{getKeywords.Rdoc}
\alias{Rdoc.getKeywords}
\alias{getKeywords,Rdoc-method}

\title{Gets the keywords defined in R with descriptions}

\description{
  Gets the keywords defined in R with descriptions.
}

\usage{
## Static method (use this):
## Rdoc$getKeywords(fullInfo=FALSE, ...)

## Don't use the below:
\method{getKeywords}{Rdoc}(this, fullInfo=FALSE, ...)
}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{character}} string.
}

\author{Henrik Bengtsson}

\seealso{
  \code{\link[R.oo:setManPath.Rdoc]{*setManPath}()}
  For more information see \code{\link{Rdoc}}.
}


\keyword{internal}
\keyword{methods}
\keyword{documentation}
