%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  
% 
% on Tue May 03 13:28:06 2005.
% 
% Generator was the Rdoc class, which is part of the R.oo package written
% by Henrik Bengtsson, 2001-2004.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%



\name{getKnownSubclasses.classRepresentation}
\alias{getKnownSubclasses.classRepresentation}
\alias{classRepresentation.getKnownSubclasses}
\alias{getKnownSubclasses.classRepresentation}
\alias{getKnownSubclasses,classRepresentation-method}
\keyword{methods}
\keyword{internal}


\title{Gets the known subclasses}

\description{
  Gets the known subclasses.
}

\usage{\method{getKnownSubclasses}{classRepresentation}(this, ...)}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{character}} string.
}

\author{Henrik Bengtsson \url{http://www.braju.com/R/}}

\seealso{
  \code{\link[R.oo:setNameFormat.classRepresentation]{*setNameFormat}()}
}

\keyword{documentation}

