%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  classRepresentation.misc.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%



\name{getKnownSubclasses.classRepresentation}
\alias{getKnownSubclasses.classRepresentation}


\title{Gets the known subclasses}

\description{
  Gets the known subclasses.
}

\usage{
\method{getKnownSubclasses}{classRepresentation}(this, ...)
}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{character}} string.
}

\author{Henrik Bengtsson}


\keyword{methods}
\keyword{internal}
\keyword{documentation}
