%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  
% 
% on Tue Feb 15 13:45:05 2005.
% 
% Generator was the Rdoc class, which is part of the R.oo package written
% by Henrik Bengtsson, 2001-2004.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getLastException.Exception}
\alias{getLastException.Exception}
\alias{Exception.getLastException}
\alias{getLastException.Exception}
\alias{getLastException,Exception-method}
\keyword{methods}
\keyword{internal}


\title{Static method to get the last Exception thrown}

\description{
 Static method to get the last Exception instanciated.
}

\synopsis{getLastException.Exception(this, ...)}
\usage{Exception$getLastException(...)}

\arguments{
  \item{...}{Not used.}
}

\value{
  Returns an \code{\link{Exception}} object.
}

\examples{\dontrun{For a complete example see help(Exception).}}

\seealso{
  For more information see \code{\link{Exception}}.
  See also \code{\link[base]{tryCatch}}().
}

\author{Henrik Bengtsson \url{http://www.braju.com/R/}}

\keyword{programming}
\keyword{methods}
\keyword{error}
