%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  Exception.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getLastException.Exception}
\alias{getLastException.Exception}
\alias{Exception.getLastException}
\alias{getLastException.Exception}
\alias{getLastException,Exception-method}


\title{Static method to get the last Exception thrown}

\description{
 Static method to get the last Exception instanciated.
}

\synopsis{getLastException.Exception(this, ...)}
\usage{Exception$getLastException(...)}

\arguments{
  \item{...}{Not used.}
}

\value{
  Returns an \code{\link{Exception}} object.
}

\examples{\dontrun{For a complete example see help(Exception).}}

\seealso{
  For more information see \code{\link{Exception}}.
  See also \code{\link[base:conditions]{tryCatch}()}.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\keyword{programming}
\keyword{methods}
\keyword{error}
\keyword{internal}
\keyword{methods}
