%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  Rdoc.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getManPath.Rdoc}
\alias{getManPath.Rdoc}
\alias{Rdoc.getManPath}
\alias{getManPath.Rdoc}
\alias{getManPath,Rdoc-method}


\title{Gets the path to the directory where the Rd files will be saved}

\description{
  Gets the path to the directory where the Rd files will be saved.
}

\usage{
## Static method (use this):
## Rdoc$getManPath(...)

## Don't use the below:
\method{getManPath}{Rdoc}(this, ...)
}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{character}} string.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  \code{\link[R.oo:setManPath.Rdoc]{*setManPath}()}
  For more information see \code{\link{Rdoc}}.
}


\keyword{internal}
\keyword{methods}
\keyword{documentation}
