%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  
% 
% on Tue May 03 13:27:36 2005.
% 
% Generator was the Rdoc class, which is part of the R.oo package written
% by Henrik Bengtsson, 2001-2004.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getName.Class}
\alias{getName.Class}
\alias{Class.getName}
\alias{getName.Class}
\alias{getName,Class-method}
\keyword{methods}
\keyword{internal}


\title{Gets the name of the class}

\description{
 Gets the name of the class.
}

\usage{\method{getName}{Class}(this, ...)}

\arguments{
  \item{...}{Not used.}
}

\value{
  Returns a \code{\link[base]{character}} string.
}

\examples{
  print(getName(Object))   # "Object"
  print(getName(Class))    # "Class"
}

\author{Henrik Bengtsson \url{http://www.braju.com/R/}}

\seealso{
  For more information see \code{\link{Class}}.
}

\keyword{programming}

\keyword{methods}

