%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  
% 
% on Tue May 03 13:28:57 2005.
% 
% Generator was the Rdoc class, which is part of the R.oo package written
% by Henrik Bengtsson, 2001-2004.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getNameFormat.Rdoc}
\alias{getNameFormat.Rdoc}
\alias{Rdoc.getNameFormat}
\alias{getNameFormat.Rdoc}
\alias{getNameFormat,Rdoc-method}
\keyword{methods}
\keyword{internal}


\title{Gets the current name format}

\description{
  Gets the current name format.
}

\synopsis{getNameFormat.Rdoc(static, ...)}
\usage{Rdoc$getNameFormat(...)}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{character}} string.
}

\author{Henrik Bengtsson \url{http://www.braju.com/R/}}

\seealso{
  \code{\link[R.oo:setNameFormat.Rdoc]{*setNameFormat}()}
  For more information see \code{\link{Rdoc}}.
}

\keyword{documentation}

