%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  Rdoc.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getNameFormat.Rdoc}
\alias{getNameFormat.Rdoc}
\alias{Rdoc.getNameFormat}
\alias{getNameFormat.Rdoc}
\alias{getNameFormat,Rdoc-method}


\title{Gets the current name format}

\description{
  Gets the current name format.
}

\usage{
## Static method (use this):
## Rdoc$getNameFormat(...)

## Don't use the below:
\method{getNameFormat}{Rdoc}(static, ...)
}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{character}} string.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  \code{\link[R.oo:setNameFormat.Rdoc]{*setNameFormat}()}
  For more information see \code{\link{Rdoc}}.
}


\keyword{internal}
\keyword{methods}
\keyword{documentation}
