%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  InternalErrorException.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getPackage.InternalErrorException}
\alias{getPackage.InternalErrorException}
\alias{InternalErrorException.getPackage}
\alias{getPackage,InternalErrorException-method}

\title{Gets the suspicious package likely to contain an error}

\description{
 Gets the suspicious package likely to contain an error.
}

\usage{
\method{getPackage}{InternalErrorException}(this, ...)
}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link{Package}} object.
}

\author{Henrik Bengtsson}

\seealso{
  For more information see \code{\link{InternalErrorException}}.
}

\keyword{programming}
\keyword{methods}
\keyword{error}
\keyword{internal}
\keyword{methods}
