%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  RccViolationException.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getRccUrl.RccViolationException}
\alias{getRccUrl.RccViolationException}
\alias{RccViolationException.getRccUrl}
\alias{getRccUrl.RccViolationException}
\alias{getRccUrl,RccViolationException-method}


\title{Static method to get a URL where the RCC can be found}

\description{
 Static method to get a URL where one can find details about the
 R Code Convention (RCC).
 Currently the URL is \url{http://aroma-project.org/developers/RCC/}.
}

\synopsis{getRccUrl.RccViolationException(this, ...)}
\usage{RccViolationException$getRccUrl(...)}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{character}} string.
}

\examples{\dontrun{For a complete example see help(RccViolationException).}}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{RccViolationException}}.
}

\keyword{programming}
\keyword{methods}
\keyword{error}
\keyword{internal}
\keyword{methods}
