%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  Class.misc.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%



\name{getRdDeclaration.Class}
\alias{getRdDeclaration.Class}
\alias{Class.getRdDeclaration}
\alias{getRdDeclaration,Class-method}

\title{Gets the class declaraction in Rd format}

\description{
  Gets the class declaraction in Rd format.
}

\usage{
\method{getRdDeclaration}{Class}(this, ...)
}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{character}} string.
}

\author{Henrik Bengtsson}

\seealso{
  For more information see \code{\link{Class}}.
}


\keyword{internal}
\keyword{methods}
\keyword{documentation}
