%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  
% 
% on Tue May 03 13:28:05 2005.
% 
% Generator was the Rdoc class, which is part of the R.oo package written
% by Henrik Bengtsson, 2001-2004.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getRdHierarchy.Class}
\alias{getRdHierarchy.Class}
\alias{Class.getRdHierarchy}
\alias{getRdHierarchy.Class}
\alias{getRdHierarchy,Class-method}
\keyword{methods}
\keyword{internal}


\title{Gets the class hierarchy in Rd format}

\description{
  Gets the class hierarchy in Rd format.
}

\usage{\method{getRdHierarchy}{Class}(this, ...)}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{character}} string.
}

\author{Henrik Bengtsson \url{http://www.braju.com/R/}}

\seealso{
  \code{\link[R.oo:setNameFormat.Class]{*setNameFormat}()}
  For more information see \code{\link{Class}}.
}

\keyword{documentation}

