%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  Class.misc.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getRdHierarchy.Class}
\alias{getRdHierarchy.Class}
\alias{Class.getRdHierarchy}
\alias{getRdHierarchy.Class}
\alias{getRdHierarchy,Class-method}


\title{Gets the class hierarchy in Rd format}

\description{
  Gets the class hierarchy in Rd format.
}

\usage{\method{getRdHierarchy}{Class}(this, ...)}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{character}} string.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  \code{\link[R.oo:setNameFormat.Class]{*setNameFormat}()}
  For more information see \code{\link{Class}}.
}


\keyword{internal}
\keyword{methods}
\keyword{documentation}
