%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  
% 
% on Mon Apr 03 17:20:13 2006.
% 
% Generator was the Rdoc class, which is part of the R.oo package written
% by Henrik Bengtsson, 2001-2006.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getRdHierarchy.classRepresentation}
\alias{getRdHierarchy.classRepresentation}
\alias{getRdHierarchy.classRepresentation}



\title{Gets the class hierarchy in Rd format}

\description{
  Gets the class hierarchy in Rd format.
}

\usage{\method{getRdHierarchy}{classRepresentation}(this, ...)}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{character}} string.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
}


\keyword{methods}
\keyword{internal}
\keyword{documentation}
