%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  Class.misc.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getRdMethods.Class}
\alias{getRdMethods.Class}
\alias{Class.getRdMethods}
\alias{getRdMethods,Class-method}

\title{Gets the methods of a class in Rd format}

\description{
  Gets the methods of a class in Rd format.
}

\usage{
\method{getRdMethods}{Class}(class, visibilities=c("private", "protected", "public"), ...)
}

\arguments{
 \item{visibilities}{A \code{\link[base]{character}} string specifying what types of methods
    to return.}
 \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{character}} string.
}

\author{Henrik Bengtsson}

\seealso{
  For more information see \code{\link{Class}}.
}


\keyword{internal}
\keyword{methods}
\keyword{documentation}
