%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  
% 
% on Tue Feb 15 13:44:59 2005.
% 
% Generator was the Rdoc class, which is part of the R.oo package written
% by Henrik Bengtsson, 2001-2004.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getRdMethods.classRepresentation}
\alias{getRdMethods.classRepresentation}
\alias{classRepresentation.getRdMethods}
\alias{getRdMethods.classRepresentation}
\alias{getRdMethods,classRepresentation-method}
\keyword{methods}
\keyword{internal}


\title{Gets the methods in Rd format}

\description{
  Gets the methods in Rd format.
}

\usage{\method{getRdMethods}{classRepresentation}(class, visibility=c("public", "protected", "private"), trial=FALSE, ...)}

\arguments{
  \item{visibility}{-}
  \item{trial}{-}
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{character}} string.
}

\author{Henrik Bengtsson \url{http://www.braju.com/R/}}

\seealso{
}

\keyword{documentation}

