%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  classRepresentation.misc.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getRdMethods.classRepresentation}
\alias{getRdMethods.classRepresentation}
\alias{getRdMethods.classRepresentation}



\title{Gets the methods in Rd format}

\description{
  Gets the methods in Rd format.
}

\usage{
\method{getRdMethods}{classRepresentation}(class, visibility=c("public", "protected", "private"), trial=FALSE, ...)
}

\arguments{
  \item{visibility}{-}
  \item{trial}{-}
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{character}} string.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}


\keyword{methods}
\keyword{internal}
\keyword{documentation}
