%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  
% 
% on Tue Feb 15 13:45:39 2005.
% 
% Generator was the Rdoc class, which is part of the R.oo package written
% by Henrik Bengtsson, 2001-2004.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getSource.RdocException}
\alias{getSource.RdocException}
\alias{RdocException.getSource}
\alias{getSource.RdocException}
\alias{getSource,RdocException-method}
\keyword{methods}
\keyword{internal}


\title{Gets the source of the exception}

\description{
 .
}

\usage{\method{getSource}{RdocException}(this, ...)}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns the source.
}

\author{Henrik Bengtsson \url{http://www.braju.com/R/}}

\seealso{
  For more information see \code{\link{RdocException}}.
}

\keyword{programming}
\keyword{methods}
\keyword{error}
