%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  Exception.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getStackTraceString.Exception}
\alias{getStackTraceString.Exception}
\alias{Exception.getStackTraceString}
\alias{getStackTraceString,Exception-method}

\title{Gets the stack trace as a string}

\description{
 .
}

\usage{
\method{getStackTraceString}{Exception}(this, ..., details=TRUE)
}

\arguments{
  \item{...}{Not used.}
}

\value{
  Returns a \code{\link[base]{character}} string.
}

\seealso{
  \code{\link[R.oo:getStackTrace.Exception]{*getStackTrace}()}.
  For more information see \code{\link{Exception}}.
}

\author{Henrik Bengtsson}

\keyword{programming}
\keyword{methods}
\keyword{error}
\keyword{internal}
\keyword{methods}
