%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  060.Class.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getStaticInstance.Class}
\alias{getStaticInstance.Class}
\alias{Class.getStaticInstance}
\alias{getStaticInstance.Class}
\alias{getStaticInstance,Class-method}


\title{Gets the static instance of this class}

\description{
  Gets the static instance of this class.
}

\usage{\method{getStaticInstance}{Class}(this, ...)}

\arguments{
  \item{...}{Not used.}
}

\value{
  Returns a reference to an \code{\link{Object}}.
}

\examples{
  obj <- getStaticInstance(Object)
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{Class}}.
}



\keyword{internal}
\keyword{methods}
\keyword{programming}
