%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  
% 
% on Mon Jul 18 15:07:42 2005.
% 
% Generator was the Rdoc class, which is part of the R.oo package written
% by Henrik Bengtsson, 2001-2004.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getSuperclasses.classRepresentation}
\alias{getSuperclasses.classRepresentation}
\alias{getSuperclasses.classRepresentation}



\title{Gets the superclasses}

\description{
  Gets the superclasses.
}

\usage{\method{getSuperclasses}{classRepresentation}(this, ...)}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{character}} string.
}

\author{Henrik Bengtsson \url{http://www.braju.com/R/}}

\seealso{
  \code{\link[R.oo:setNameFormat.classRepresentation]{*setNameFormat}()}
}


\keyword{methods}
\keyword{internal}
\keyword{documentation}
