%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  classRepresentation.misc.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getSuperclasses.classRepresentation}
\alias{getSuperclasses.classRepresentation}


\title{Gets the superclasses}

\description{
  Gets the superclasses.
}

\usage{
\method{getSuperclasses}{classRepresentation}(this, ...)
}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{character}} string.
}

\author{Henrik Bengtsson}


\keyword{methods}
\keyword{internal}
\keyword{documentation}
