%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  Package.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getVersion.Package}
\alias{getVersion.Package}
\alias{Package.getVersion}
\alias{getVersion.Package}
\alias{getVersion,Package-method}


\title{Gets the version of this package}

\description{
  Gets the version of this package.
}

\usage{
\method{getVersion}{Package}(this, ...)
}

\arguments{
  \item{...}{Not used.}
}

\value{
  Returns a \code{\link[base]{character}} string.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{Package}}.
  \code{\link[R.oo:isOlderThan.Package]{*isOlderThan}()}.
}
\keyword{internal}
\keyword{methods}
