%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  
% 
% on Tue May 03 13:28:17 2005.
% 
% Generator was the Rdoc class, which is part of the R.oo package written
% by Henrik Bengtsson, 2001-2004.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getWhen.Exception}
\alias{getWhen.Exception}
\alias{Exception.getWhen}
\alias{getWhen.Exception}
\alias{getWhen,Exception-method}
\keyword{methods}
\keyword{internal}


\title{Gets the time when the Exception was created}

\description{
 Gets the time, as a POSIX object, when the Exception was created.
}

\usage{\method{getWhen}{Exception}(this, ...)}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a POSIX time object.
}

\examples{\dontrun{For a complete example see help(Exception).}}

\author{Henrik Bengtsson \url{http://www.braju.com/R/}}

\seealso{
  For more information see \code{\link{Exception}}.
}

\keyword{programming}
\keyword{methods}
\keyword{error}
