%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  Exception.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getWhen.Exception}
\alias{getWhen.Exception}
\alias{Exception.getWhen}
\alias{getWhen,Exception-method}

\title{Gets the time when the Exception was created}

\description{
 Gets the time, as a POSIX object, when the Exception was created.
}

\usage{
\method{getWhen}{Exception}(this, ...)
}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a POSIX time object.
}

\examples{\dontrun{For a complete example see help(Exception).}}

\author{Henrik Bengtsson}

\seealso{
  For more information see \code{\link{Exception}}.
}

\keyword{programming}
\keyword{methods}
\keyword{error}
\keyword{internal}
\keyword{methods}
