%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  Rdoc.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{hierarchy.Rdoc}
\alias{hierarchy.Rdoc}
\alias{Rdoc.hierarchy}
\alias{hierarchy.Rdoc}
\alias{hierarchy,Rdoc-method}


\title{Gets the class hierarchy}

\description{
  Gets the class hierarchy.
}

\usage{
## Static method (use this):
## Rdoc$hierarchy(class, ...)

## Don't use the below:
\method{hierarchy}{Rdoc}(this, class, ...)
}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{character}} string.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{Rdoc}}.
}


\keyword{internal}
\keyword{methods}
\keyword{documentation}
