%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  Interface.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{Interface}
\docType{class}
\alias{Interface}


 \title{The Interface class}

 \description{
  Package:  R.oo \cr
\bold{Class Interface}\cr

\code{+--}\code{Interface}\cr

\bold{Directly known subclasses:}\cr
\cr

public class \bold{Interface}\cr



  This class represents a special set of classes whose purpose is to
  provide methods (but not fields) shared by multiple different classes.
 }

 \usage{
Interface(core=NA, ...)
}

 \arguments{
   \item{core}{The core value.}
   \item{...}{Not used.}
 }

 \section{Methods}{
  \bold{Methods:}\cr
\tabular{rll}{
 \tab \code{\link[R.oo:extend.Interface]{extend}} \tab Extends another Interface class.\cr
 \tab \code{\link[R.oo:print.Interface]{print}} \tab Prints an Interface.\cr
 \tab \code{\link[R.oo:uses.Interface]{uses}} \tab Specifies that an object uses this Interface.\cr
}



 }

 \author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\keyword{classes}
\keyword{internal}
