%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  Rdoc.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{isKeyword.Rdoc}
\alias{isKeyword.Rdoc}
\alias{Rdoc.isKeyword}
\alias{isKeyword.Rdoc}
\alias{isKeyword,Rdoc-method}


\title{Checks if a word is a Rd keyword}

\description{
  Checks if a word is a Rd keyword.
}

\synopsis{isKeyword.Rdoc(this, word, ...)}
\usage{Rdoc$isKeyword(word, ...)}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{logical}}.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  \code{\link[R.oo:getKeywords.Rdoc]{*getKeywords}()}
  For more information see \code{\link{Rdoc}}.
}


\keyword{internal}
\keyword{methods}
\keyword{documentation}
