%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  BasicObject.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{isReferable.BasicObject}
\alias{isReferable.BasicObject}
\alias{BasicObject.isReferable}
\alias{isReferable,BasicObject-method}

\title{Checks if the object is referable or not}

\description{
 Checks if the object is referable or not.
}

\usage{
\method{isReferable}{BasicObject}(this, ...)
}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{logical}} value, which by default is \code{\link[base:logical]{TRUE}} for all
 \code{\link{BasicObject}}'s.
}

\author{Henrik Bengtsson}

\seealso{
  For more information see \code{\link{BasicObject}}.
}

\keyword{programming}
\keyword{methods}
\keyword{internal}
\keyword{methods}
