%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  060.Class.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{print.Class}
\alias{print.Class}
\alias{Class.print}
\alias{print,Class-method}

\title{Prints detailed information about the class and its fields and methods}

\description{
 Prints detailed information about the class and its fields and methods.
}

\usage{
\method{print}{Class}(x, ...)
}

\arguments{
  \item{...}{Arguments passed to \code{\link[R.oo:getDetails.Class]{*getDetails}()}.}
}

\value{
  Returns nothing.
}

\examples{
  print(Object)
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  \code{\link[R.oo:getDetails.Class]{*getDetails}()}
  For more information see \code{\link{Class}}.
}



\keyword{internal}
\keyword{methods}
\keyword{programming}
