%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  Exception.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{print.Exception}
\alias{print.Exception}
\alias{Exception.print}
\alias{print,Exception-method}

\title{Prints the Exception}

\description{
 . By default the \code{as.character()} representation plus
 the stack trace is printed.
}

\usage{
\method{print}{Exception}(x, ...)
}

\arguments{
  \item{...}{Not used.}
}

\value{Returns nothing.}

\examples{\dontrun{For a complete example see help(Exception).}}

\author{Henrik Bengtsson}

\seealso{
  \code{\link[R.oo:as.character.Exception]{*as.character}()}.
  \code{\link[R.oo:getStackTrace.Exception]{*getStackTrace}()}.
  \code{\link[R.oo:printStackTrace.Exception]{*printStackTrace}()}.
  For more information see \code{\link{Exception}}.
}

\keyword{programming}
\keyword{methods}
\keyword{error}
\keyword{internal}
\keyword{methods}
