%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  
% 
% on Tue May 03 13:27:18 2005.
% 
% Generator was the Rdoc class, which is part of the R.oo package written
% by Henrik Bengtsson, 2001-2004.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{print.Object}
\alias{print.Object}
\alias{Object.print}
\alias{print.Object}
\alias{print,Object-method}
\keyword{methods}
\keyword{internal}


\title{Prints an Object}

\description{
 For all objects of class \code{\link{Object}}, this method will print the
 value of \code{as.character()} of the object. Note that this function is
 not called if the argument is not an object of class \code{\link[R.oo]{Object}}.
}

\usage{\method{print}{Object}(x, ...)}

\examples{\dontrun{For a complete example see help(Object).}}

\author{Henrik Bengtsson \url{http://www.braju.com/R/}}

\seealso{
  \code{\link[base]{print.default}}()
  For more information see \code{\link{Object}}.
}

\keyword{programming}
\keyword{methods}
