%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  050.Object.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{print.Object}
\alias{print.Object}
\alias{Object.print}
\alias{print,Object-method}

\title{Prints an Object}

\description{
 For all objects of class \code{\link{Object}}, this method will print the
 value of \code{as.character()} of the object. Note that this function is
 not called if the argument is not an object of class \code{\link[R.oo]{Object}}.
}

\usage{
\method{print}{Object}(x, ...)
}

\examples{\dontrun{For a complete example see help(Object).}}

\author{Henrik Bengtsson}

\seealso{
  \code{\link[base]{print.default}}()
  For more information see \code{\link{Object}}.
}

\keyword{programming}
\keyword{methods}
\keyword{internal}
\keyword{methods}
