%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  Exception.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{printStackTrace.Exception}
\alias{printStackTrace.Exception}
\alias{Exception.printStackTrace}
\alias{printStackTrace,Exception-method}

\title{Prints the stack trace saved when the exception was created}

\description{
 .
}

\usage{
\method{printStackTrace}{Exception}(this, ...)
}

\arguments{
  \item{...}{Not used.}
}

\value{
  Returns nothing.
}

\examples{\dontrun{For a complete example see help(Exception).}}

\seealso{
  \code{\link[R.oo:getStackTrace.Exception]{*getStackTrace}()}.
  \code{\link[base:conditions]{tryCatch}()}.
  For more information see \code{\link{Exception}}.
}

\author{Henrik Bengtsson}

\keyword{programming}
\keyword{methods}
\keyword{error}
\keyword{internal}
\keyword{methods}
