%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  001.R.KEYWORDS.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{R.KEYWORDS}
\alias{R.KEYWORDS}

\title{Reserved words in R not to be used for object names}

\description{
 Reserved words in R not to be used for object names. \code{R.KEYWORDS} is a \code{\link[base]{character}} \code{\link[base]{vector}} of all reserved
 words in \R according to [1].
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\references{
 [1] Section "Reserved words", R Language Definition, version 2.4.0
     (2006-07-25) DRAFT.
}



\keyword{programming}
\keyword{internal}
