%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  
% 
% on Tue May 03 13:28:34 2005.
% 
% Generator was the Rdoc class, which is part of the R.oo package written
% by Henrik Bengtsson, 2001-2004.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{showContents.Package}
\alias{showContents.Package}
\alias{Package.showContents}
\alias{showContents.Package}
\alias{showContents,Package-method}
\keyword{methods}
\keyword{internal}


\title{Show the CONTENTS file of this package}

\description{
  Show the CONTENTS file of this package.
  If the \code{CONTENTS} file does not exist, an exception is thrown.
}

\usage{\method{showContents}{Package}(this, ...)}

\arguments{
  \item{...}{Not used.}
}

\value{
  Returns nothing.
}

\author{Henrik Bengtsson \url{http://www.braju.com/R/}}

\seealso{
  For more information see \code{\link{Package}}.
}
