%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  Package.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{showDescriptionFile.Package}
\alias{showDescriptionFile.Package}
\alias{Package.showDescriptionFile}
\alias{showDescriptionFile,Package-method}

\title{Show the DESCRIPTION file of this package}

\description{
  Show the DESCRIPTION file of this package.
  If the \code{DESCRIPTION} file does not exist, an exception is thrown.
}

\usage{
\method{showDescriptionFile}{Package}(this, ...)
}

\arguments{
  \item{...}{Not used.}
}

\value{
  Returns nothing.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{Package}}.
}
\keyword{internal}
\keyword{methods}
