%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  
% 
% on Tue Feb 15 13:45:24 2005.
% 
% Generator was the Rdoc class, which is part of the R.oo package written
% by Henrik Bengtsson, 2001-2004.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{showHistory.Package}
\alias{showHistory.Package}
\alias{Package.showHistory}
\alias{showHistory.Package}
\alias{showHistory,Package-method}
\keyword{methods}
\keyword{internal}


\title{Show the HISTORY file of this package}

\description{
  Show the HISTORY file of this package. See also \code{\link[R.oo:getHistory.Package]{*getHistory}()}.
  If the \code{HISTORY} file does not exist, an exception is thrown.
}

\usage{\method{showHistory}{Package}(this, ...)}

\arguments{
  \item{...}{Not used.}
}

\value{
  Returns nothing.
}

\author{Henrik Bengtsson \url{http://www.braju.com/R/}}

\seealso{
  For more information see \code{\link{Package}}.
}
