%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  
% 
% on Tue May 03 13:28:49 2005.
% 
% Generator was the Rdoc class, which is part of the R.oo package written
% by Henrik Bengtsson, 2001-2004.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{showHowToCite.Package}
\alias{showHowToCite.Package}
\alias{Package.showHowToCite}
\alias{showHowToCite.Package}
\alias{showHowToCite,Package-method}
\keyword{methods}
\keyword{internal}


\title{Show the HOWTOCITE file of this package}

\description{
  Show the HOWTOCITE file of this package. See also \code{\link[R.oo:getHowToCite.Package]{*getHowToCite}()}.
  If the \code{HOWTOCITE} file does not exist, an exception is thrown.
}

\usage{\method{showHowToCite}{Package}(this, ...)}

\arguments{
  \item{...}{Not used.}
}

\value{
  Returns nothing.
}

\author{Henrik Bengtsson \url{http://www.braju.com/R/}}

\seealso{
  For more information see \code{\link{Package}}.
}
