%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  Package.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{showHowToCite.Package}
\alias{showHowToCite.Package}
\alias{Package.showHowToCite}
\alias{showHowToCite.Package}
\alias{showHowToCite,Package-method}


\title{Show the HOWTOCITE file of this package}

\description{
  Show the HOWTOCITE file of this package. See also \code{\link[R.oo:getHowToCite.Package]{*getHowToCite}()}.
  If the \code{HOWTOCITE} file does not exist, an exception is thrown.
}

\usage{
\method{showHowToCite}{Package}(this, ...)
}

\arguments{
  \item{...}{Not used.}
}

\value{
  Returns nothing.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{Package}}.
}
\keyword{internal}
\keyword{methods}
