%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  Package.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{startupMessage.Package}
\alias{startupMessage.Package}
\alias{Package.startupMessage}
\alias{startupMessage,Package-method}

\title{Generates a 'package successfully loaded' package startup message}

\description{
  Generates a 'package successfully loaded' package startup message.
}

\usage{
\method{startupMessage}{Package}(this, ...)
}

\arguments{
  \item{...}{Additional arguments passed to
             \code{\link[base]{packageStartupMessage}}().}
}

\value{
  Returns nothing.
}

\author{Henrik Bengtsson}

\seealso{
  \code{\link[base]{packageStartupMessage}}().

  For more information see \code{\link{Package}}.
}
\keyword{internal}
\keyword{methods}
