%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  
% 
% on Mon Jul 18 15:07:49 2005.
% 
% Generator was the Rdoc class, which is part of the R.oo package written
% by Henrik Bengtsson, 2001-2004.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{throw.Exception}
\alias{throw.Exception}
\alias{Exception.throw}
\alias{throw.Exception}
\alias{throw,Exception-method}


\title{Throws an Exception that can be caught}

\description{
 Throws an Exception that can be caught by \code{tryCatch()}.
}

\usage{\method{throw}{Exception}(this, ...)}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns nothing.
}

\examples{\dontrun{For a complete example see help(Exception).}}

\author{Henrik Bengtsson \url{http://www.braju.com/R/}}

\seealso{
  For more information see \code{\link{Exception}}.
  See also \code{\link[base]{tryCatch}}().
}

\keyword{programming}
\keyword{methods}
\keyword{error}
\keyword{internal}
\keyword{methods}
