%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  
% 
% on Tue Feb 15 13:45:40 2005.
% 
% Generator was the Rdoc class, which is part of the R.oo package written
% by Henrik Bengtsson, 2001-2004.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{trim}
\alias{trim.default}
\alias{trim}

\title{Converts to a string and removes leading and trailing whitespace}

\description{
 Converts to a string and removes leading and trailing whitespace.
}

\usage{\method{trim}{default}(object, ...)}

\arguments{
 \item{object}{A \code{\link[base]{vector}} of \R objects to be trimmed.}
 \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{vector}} of \code{\link[base]{character}} strings.
}

\author{Henrik Bengtsson \url{http://www.braju.com/R/}}

\keyword{character}
