%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  trim.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{trim}
\alias{trim.default}
\alias{trim}

\title{Converts to a string and removes leading and trailing whitespace}

\description{
 Converts to a string and removes leading and trailing whitespace.
}

\usage{\method{trim}{default}(object, ...)}

\arguments{
 \item{object}{A \code{\link[base]{vector}} of \R objects to be trimmed.}
 \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{vector}} of \code{\link[base]{character}} strings.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\keyword{character}

